widget carousel

+++ backend/controllers/WidgetCarouselController.php
@@ -2,32 +2,46 @@
 
 namespace backend\controllers;
 
-use backend\models\search\WidgetCarouselItemSearch;
 use Yii;
 use common\models\WidgetCarousel;
-use common\models\WidgetCarouselItem;
-use backend\models\search\WidgetCarouselSearch;
-use yii\web\Controller;
+use common\models\search\WidgetCarouselSearch;
+use common\components\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-use yii\web\Response;
-use yii\web\UploadedFile;
+use yii\helpers\Json;
+use yii\helpers\Url;
+use yii\helpers\ArrayHelper;
+use kartik\grid\EditableColumnAction;
 
 /**
  * WidgetCarouselController implements the CRUD actions for WidgetCarousel model.
  */
 class WidgetCarouselController extends Controller
 {
+    /**
+     * @inheritdoc
+     */
     public function behaviors()
     {
         return [
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
-                    'delete' => ['post']
+                    'delete' => ['POST'],
                 ],
             ],
         ];
+    }
+
+    public function actions()
+    {
+        return ArrayHelper::merge(parent::actions(), [
+            'edit' => [
+                'class' => EditableColumnAction::className(),
+                'modelClass' => WidgetCarousel::className(),
+                'showModelErrors' => true,
+            ]
+        ]);
     }
 
     /**
@@ -45,6 +59,25 @@ class WidgetCarouselController extends Controller
         ]);
     }
 
+    /**
+     * Displays a single WidgetCarousel model.
+     * @param integer $id
+     * @return mixed
+     */
+    public function actionView($id)
+    {
+        $model=$this->findModel($id);
+
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if (!Yii::$app->user->can('/' . \common\components\helpers\Url::normalizeRoute('update'))) {
+                throw new HttpException(403);
+            }
+            Yii::$app->session->setFlash('kv-detail-success', 'Saved record successfully');
+            return $this->redirect(['view', 'id' => $model->id]);
+        } else {
+            return $this->render('view', ['model'=>$model]);
+        }
+    }
 
     /**
      * Creates a new WidgetCarousel model.
@@ -56,7 +89,7 @@ class WidgetCarouselController extends Controller
         $model = new WidgetCarousel();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['update', 'id' => $model->id]);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('create', [
                 'model' => $model,
@@ -74,16 +107,11 @@ class WidgetCarouselController extends Controller
     {
         $model = $this->findModel($id);
 
-        $searchModel = new WidgetCarouselItemSearch();
-        $carouselItemsProvider = $searchModel->search([]);
-        $carouselItemsProvider->query->andWhere(['carousel_id'=>$model->id]);
-
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index']);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('update', [
                 'model' => $model,
-                'carouselItemsProvider'=>$carouselItemsProvider
             ]);
         }
     }
@@ -94,11 +122,31 @@ class WidgetCarouselController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+    public function actionDelete($id) {
+        $post = Yii::$app->request->post();
+        if (Yii::$app->request->isAjax && isset($post['j-delete'])) {
+            if ($this->findModel($id)->delete()) {
+                echo Json::encode([
+                    'success' => true,
+                    'messages' => [
+                    'kv-detail-info' => 'Successfully deleted. <a href="' .
+                                                Url::to(['index']) . '" class="btn btn-sm btn-info">' .
+                        '<i class="glyphicon glyphicon-hand-right"></i>  Click here</a> to proceed.'
+                    ]
+                ]);
+            } else {
+                echo Json::encode([
+                    'success' => false,
+                    'messages' => [
+                        'kv-detail-error' => 'Cannot delete'
+                    ]
+                ]);
+            }
+            return;
+        } else {
+            $this->findModel($id)->delete();
+            $this->redirect(['index', 'returned'=>true]);
+        }
     }
 
     /**
